feat(api-requester): add method to get random bad guy image
Implement a new method `getBadGuyImg` in the `ApiRequester` class to fetch a random
image of a specified bad guy type. This method sends a request to the LiuLiApiwith the bad guy name and retrieves the image URL. The image is then downloaded
and returned as a byte array. Additionally, introduced the `BadGuyType` enumeration
to manage different types of bad guys and provide a random selection mechanism.